Switch randString to use crypto random and put it with jwt stuff.
